+++ src/mwin/winlib/newedit.c
@@ -187,7 +187,7 @@ neCreate(HWND hwnd)
 	if (len < LEN_SLEDIT_BUFFER)
 		len = LEN_SLEDIT_BUFFER;
 
-	pCtrl->userdata2 = (DWORD) pSLEditData;
+	pCtrl->userdata2 = (LONG) pSLEditData;
 	pCtrl->userdata = 0;
 
 #ifdef USE_BIG5
@@ -264,7 +264,7 @@ neRecalcRows(HWND hwnd, PSLEDITDATA * ppData)
 	pSLEditData->charHeight = charH;
 	pSLEditData->cLines = nl;
 	*ppData = pSLEditData;
-	hwnd->userdata2 = (DWORD) pSLEditData;
+	hwnd->userdata2 = (LONG) pSLEditData;
 	free(pCurData);
 }
